updated userdata scripts, security groups

+++ userdata/userdata.tpl
@@ -32,17 +32,58 @@ $ad_dns_ip_1="${ad_dns_ip_1}"
 $ad_dns_ip_2="${ad_dns_ip_2}"
 Set-DNSClientServerAddress -interfaceIndex $serverAddresses.InterfaceIndex -ServerAddresses ("$ad_dns_ip_1", "$ad_dns_ip_2")
 
+# Get the instance id from ec2 meta data
+$instanceid = Invoke-RestMethod "http://169.254.169.254/latest/meta-data/instance-id"
+
+# Get the environment name and application from this instance's environment-name and application tag values
+$environmentName = Get-EC2Tag -Filter @(
+        @{
+            name="resource-id"
+            values="$instanceid"
+        }
+        @{
+            name="key"
+            values="environment-name"
+        }
+    )
+$application = Get-EC2Tag -Filter @(
+        @{
+            name="resource-id"
+            values="$instanceid"
+        }
+        @{
+            name="key"
+            values="application"
+        }
+    )
+
+Write-Output "environmentName: $($environmentName.Value)"
+Write-Output "application:     $($application.Value)"
+
+$ad_admin_username_SSMPath = "/" + $environmentName.Value + "/" + $application.Value + "/mis-activedirectory/ad/ad_admin_username"
+$ad_admin_password_SSMPath = "/" + $environmentName.Value + "/" + $application.Value + "/mis-activedirectory/ad/ad_admin_password"
+$ad_admin_username = Get-SSMParameter -Name $ad_admin_username_SSMPath -WithDecryption $true
+$ad_admin_password = Get-SSMParameter -Name $ad_admin_password_SSMPath -WithDecryption $true
+$secpasswd = ConvertTo-SecureString $ad_admin_password.Value -AsPlainText -Force
+
+$domainusername = "${ad_domain_name}" + "\" + $ad_admin_username.Value
+$domainusername
+$domaincreds = New-Object System.Management.Automation.PSCredential ($domainusername, $secpasswd) 
+
+# Now add the computer to the domain if its not already in the domain
+$domain=(Get-WmiObject win32_computersystem).Domain
+Write-Host $domain
+
+if( $domain -eq "WORKGROUP") {
+    Add-Computer -DomainName "${ad_domain_name}" -Credential $domaincreds
+    Restart-Computer -Force
+}
+
+# Now map the FSx filesystem (as we're now on the domain)
 # Windows 2016 Allows global mapping as 
 # see https://medium.com/@bberkayilmaz/mounting-aws-fsx-with-autoscaling-in-terraform-eee3d115d49c - scenario 3
-$secpasswd = ConvertTo-SecureString ${ad_admin_user_password} -AsPlainText -Force
-$domaincreds = New-Object System.Management.Automation.PSCredential ("${ad_domain_name}\${ad_admin_user_name}", $secpasswd)
 New-SmbGlobalMapping -RemotePath "\\${bfs_filesystem_dns_name}\share" -Credential $domaincreds -LocalPath Z:
-
-# Now add the computer to the domain
-Add-Computer -DomainName "${ad_domain_name}" -Credential $domaincreds
-Restart-Computer -Force
+Get-SmbGlobalMapping 
 
 </powershell>
 <persist>true</persist>
-
-
